refactor: incapsulate task modal
